Review: remove JSX
